+++ main.py
@@ -1,14 +1,14 @@
-
-users:list=[
-    {'name':'Beata','location':'Lublin','posts':555},
-    {'name':'Mikołaj','location':'Przasnysz','posts':200},
-    {'name':'Krzysztof','location':'Poznań','posts':100},
-    {'name':'Bartosz','location':'Ostrołęka','posts':300},
-]
+from utils.model import users
+from utils.controller import get_user_info
 
 
-def get_user_info(users_data:list)->None:
-    for user in users_data:
-        print(f'Twój znajomy {user["name"]} z miejscowości {user["location"]} opublikował {user["posts"]} postów.')
+def main():
+    get_user_info(users)
 
-get_user_info(users)
+if __name__ == '__main__':
+    main()
+
+
+
+
+
