+++ utils/model.py
@@ -1,0 +1,6 @@
+users:list=[
+    {'name':'Beata','location':'Lublin','posts':555},
+    {'name':'Mikołaj','location':'Przasnysz','posts':200},
+    {'name':'Krzysztof','location':'Poznań','posts':100},
+    {'name':'Bartosz','location':'Ostrołęka','posts':300},
+]
